A mega commit
This one brings schism up to its most recent. It has the hard drive stuff as much as I've made, it has a lot of stuff being added to libc and some testing code to get constants for libc. I've committed every file because honestly I do not know what is and is not updated anymore.